requirements.txt
google-generativeai
python-telegram-bot


# dot_env
python-dotenv

# Formating code:
flake8
black
isort